+++ totalsegmentator/libs.py
@@ -74,37 +74,54 @@ def download_url_and_unpack(url, config_dir):
 
 def download_pretrained_weights(task_id):
 
-    config_dir = Path.home() / ".totalsegmentator/nnunet/results/nnUNet/3d_fullres"
-    config_dir.mkdir(exist_ok=True, parents=True)
+    config_dir = Path.home() / ".totalsegmentator/nnunet/results/nnUNet"
+    (config_dir / "3d_fullres").mkdir(exist_ok=True, parents=True)
+    (config_dir / "2d").mkdir(exist_ok=True, parents=True)
 
     old_weights = [
         "Task223_my_test"
     ]
 
     if task_id == 251:
+        config_dir = config_dir / "3d_fullres"
         weights_path = config_dir / "Task251_TotalSegmentator_part1_organs_1139subj"
         WEIGHTS_URL = "https://zenodo.org/record/6802342/files/Task251_TotalSegmentator_part1_organs_1139subj.zip?download=1"
     elif task_id == 252:
+        config_dir = config_dir / "3d_fullres"
         weights_path = config_dir / "Task252_TotalSegmentator_part2_vertebrae_1139subj"
         WEIGHTS_URL = "https://zenodo.org/record/6802358/files/Task252_TotalSegmentator_part2_vertebrae_1139subj.zip?download=1"
     elif task_id == 253:
+        config_dir = config_dir / "3d_fullres"
         weights_path = config_dir / "Task253_TotalSegmentator_part3_cardiac_1139subj"
         WEIGHTS_URL = "https://zenodo.org/record/6802360/files/Task253_TotalSegmentator_part3_cardiac_1139subj.zip?download=1"
     elif task_id == 254:
+        config_dir = config_dir / "3d_fullres"
         weights_path = config_dir / "Task254_TotalSegmentator_part4_muscles_1139subj"
         WEIGHTS_URL = "https://zenodo.org/record/6802366/files/Task254_TotalSegmentator_part4_muscles_1139subj.zip?download=1"
     elif task_id == 255:
+        config_dir = config_dir / "3d_fullres"
         weights_path = config_dir / "Task255_TotalSegmentator_part5_ribs_1139subj"
         WEIGHTS_URL = "https://zenodo.org/record/6802452/files/Task255_TotalSegmentator_part5_ribs_1139subj.zip?download=1"
     elif task_id == 256:
+        config_dir = config_dir / "3d_fullres"
         weights_path = config_dir / "Task256_TotalSegmentator_3mm_1139subj"
         WEIGHTS_URL = "https://zenodo.org/record/6802052/files/Task256_TotalSegmentator_3mm_1139subj.zip?download=1"
     elif task_id == 258:
+        config_dir = config_dir / "3d_fullres"
         weights_path = config_dir / "Task258_lung_vessels_248subj"
         WEIGHTS_URL = "https://zenodo.org/record/7064718/files/Task258_lung_vessels_248subj.zip?download=1"
     elif task_id == 200:
+        config_dir = config_dir / "3d_fullres"
         weights_path = config_dir / "Task200_covid_challenge"
         WEIGHTS_URL = "TODO"
+    # elif task_id == 152:
+    #     config_dir = config_dir / "2d"
+    #     weights_path = config_dir / "Task152_icbbig_TN"
+    #     WEIGHTS_URL = "TODO"
+    elif task_id == 150:
+        config_dir = config_dir / "3d_fullres"
+        weights_path = config_dir / "Task150_icb_v0"
+        WEIGHTS_URL = "https://zenodo.org/record/7079161/files/Task150_icb_v0.zip?download=1"
 
     for old_weight in old_weights:
         if (config_dir / old_weight).exists():
@@ -135,6 +152,7 @@ def setup_nnunet():
     else:
         config_dir = Path.home() / ".totalsegmentator"
         (config_dir / "nnunet/results/nnUNet/3d_fullres").mkdir(exist_ok=True, parents=True)
+        (config_dir / "nnunet/results/nnUNet/2d").mkdir(exist_ok=True, parents=True)
         weights_dir = config_dir / "nnunet/results"
 
     # This variables will only be active during the python script execution. Therefore